feat: Introduce TransitionData and KernelInitialParams structures for boot ABI and refactor module organization
