+++ alkos/kernel/arch/x86_64/boot/lib/elf/elf64.cpp
@@ -1,5 +1,5 @@
 #include <string.h>
-#include <elf64.hpp>
+#include <elf/elf64.hpp>
 #include <extensions/algorithm.hpp>
 #include <extensions/bit.hpp>
 #include <extensions/debug.hpp>
