+++ alkos/kernel/arch/x86_64/boot/lib/elf/elf64.hpp
@@ -1,5 +1,5 @@
-#ifndef ALKOS_KERNEL_ARCH_X86_64_COMMON_LOADER_ALL_ELF64_HPP_
-#define ALKOS_KERNEL_ARCH_X86_64_COMMON_LOADER_ALL_ELF64_HPP_
+#ifndef ALKOS_BOOT_LIB_ELF_ELF64_HPP_
+#define ALKOS_BOOT_LIB_ELF_ELF64_HPP_
 
 #include <stdint.h>
 
@@ -97,4 +97,4 @@ std::tuple<u64, u64> GetElf64ProgramBounds(const byte* elf_start);
 
 }  // namespace elf
 
-#endif  // ALKOS_KERNEL_ARCH_X86_64_COMMON_LOADER_ALL_ELF64_HPP_
+#endif  // ALKOS_BOOT_LIB_ELF_ELF64_HPP_
