+++ alkos/kernel/arch/x86_64/boot/lib/hw/vga.hpp
