+++ alkos/kernel/arch/x86_64/boot/lib/mem/memory_manager.cpp
@@ -1,4 +1,4 @@
-#include "loader_memory_manager/loader_memory_manager.hpp"
+#include "mem/memory_manager.hpp"
 #include <assert.h>
 #include <string.h>
 #include <extensions/algorithm.hpp>
