+++ alkos/kernel/arch/x86_64/boot/lib/mem/memory_span.hpp
