+++ alkos/kernel/arch/x86_64/boot/lib/sys/constants.hpp
