+++ alkos/kernel/arch/x86_64/boot/lib/sys/terminal.cpp
@@ -1,5 +1,5 @@
 #include <drivers/serial/qemu.hpp>
-#include <terminal.hpp>
+#include <sys/terminal.hpp>
 
 void TerminalInit()
 {
